Move StatsForecast models to inherit from AbstractLocalModel
