+++ timeseries/src/autogluon/timeseries/models/local/__init__.py
@@ -1,2 +1,7 @@
 from .naive import NaiveModel, SeasonalNaiveModel
+from .statsforecast import (
+    AutoARIMAStatsForecastModel,
+    AutoETSStatsForecastModel,
+    DynamicOptimizedThetaStatsForecastModel,
+)
 from .statsmodels import ARIMAModel, ETSModel, ThetaModel
